requirements.txt
Pillow==9.3.0
PyMySQL==1.0.2
pyperclip==1.8.2
